+++ changelog.md
@@ -1,7 +1,13 @@
-- [v1.0.1 Stable](#v101-stable)
+- [v1.1.0](#v110)
+- [v1.0.1 (rejected)](#v101-rejected)
   - [backend](#backend)
 
-## v1.0.1 Stable
+## v1.1.0
+- removed youtube downloader from webstore version (requested by the chrome webstore team)
+- added twitch auto-claiming
+- now saving config locally
+
+## v1.0.1 (rejected)
 - Youtube downloader now uses a [subdomain](https://ytdl.yuji.app/)
 - Config gets cached on startup
 - Config updates periodically